Create review show page with ratings
Make models for rating and for a rating check. (A rating check would be, for
example, saying that a review is kind. By making it a model object, we allow for
extensibility in the future, i.e. if we want to add the option to evaluate a
review on additional criteria other than kind, specific, and actionable.)

Make routes for review show page. Make controller for reviews to display the ratings
associated with a review and the rating checks associated with each rating.

Make a view for review show page, which should display the content of the
review, followed by the ratings associated with it.

Also include pry in Gemfile due to its help in debugging.

Rename and reorganize integration tests.